broke up server.go to multiple files
