+++ server/server.go
@@ -1,7 +1,6 @@
 package server
 
 import (
-	"bytes"
 	"encoding/json"
 	"fmt"
 	"github.com/go-chi/chi"
@@ -10,15 +9,9 @@ import (
 	toml "github.com/pelletier/go-toml"
 	"github.com/r3labs/sse"
 	"github.com/rs/cors"
-	"github.com/shirou/gopsutil/process"
-	"github.com/spf13/viper"
 	"github.com/stellar/go/clients/horizon"
 	"log"
 	"net/http"
-	"os"
-	"os/exec"
-	"os/user"
-	"path/filepath"
 	"strings"
 	"time"
 )
@@ -142,48 +135,6 @@ func getExchanges(w http.ResponseWriter, r *http.Request) {
 	w.Write([]byte(result))
 }
 
-func configPath(id string, projectID string) string {
-	result := ""
-
-	usr, _ := user.Current()
-	configsDir := filepath.Join(usr.HomeDir, ".kelp")
-
-	// on docker the configs are located at /configs, otherwise ./configs
-	if _, err := os.Stat(configsDir); os.IsNotExist(err) {
-		configsDir = "/configs"
-	}
-
-	// get project folder, will be a param, for now just use default
-	dirName := projectID
-	if len(dirName) == 0 {
-		dirName = "default"
-	}
-
-	configsDir = filepath.Join(configsDir, dirName)
-
-	switch id {
-	case "botConf":
-		result = filepath.Join(configsDir, "trader.toml")
-		break
-	case "sell":
-		result = filepath.Join(configsDir, "sell.toml")
-		break
-	case "mirror":
-		result = filepath.Join(configsDir, "mirror.toml")
-		break
-	case "balanced":
-		result = filepath.Join(configsDir, "balanced.toml")
-		break
-	case "buysell":
-		result = filepath.Join(configsDir, "buysell.toml")
-		break
-	default:
-		break
-	}
-
-	return result
-}
-
 func getURLParam(r *http.Request, key string) string {
 	keys, ok := r.URL.Query()[key]
 
@@ -237,35 +188,6 @@ func launchDelete(w http.ResponseWriter, r *http.Request) {
 	w.Write([]byte("trade deleted"))
 }
 
-func getConfig(w http.ResponseWriter, r *http.Request) {
-	projectId := getURLParam(r, "project")
-
-	t, err := toml.TreeFromMap(configFields(projectId))
-	if err != nil {
-		log.Println(fmt.Errorf("error config file: %s \n", err))
-	}
-
-	// log.Println(t.Get("horizon_url"))
-
-	w.Write([]byte(t.String()))
-}
-
-func configFields(projectId string) map[string]interface{} {
-	configPath := configPath("botConf", projectId)
-
-	nameNoExt := filepath.Base(configPath)
-	nameNoExt = strings.TrimSuffix(nameNoExt, filepath.Ext(configPath))
-
-	viper.SetConfigName(nameNoExt)
-	viper.AddConfigPath(filepath.Dir(configPath))
-	err := viper.ReadInConfig()
-	if err != nil {
-		log.Println(fmt.Errorf("Fatal error config file: %s \n", err))
-	}
-
-	return viper.AllSettings()
-}
-
 func getOffers(w http.ResponseWriter, r *http.Request) {
 	projectId := getURLParam(r, "project")
 
@@ -288,73 +210,4 @@ func getOffers(w http.ResponseWriter, r *http.Request) {
 	js, _ := json.Marshal(offers)
 
 	w.Write(js)
-}
-
-func getProcesses(w http.ResponseWriter, r *http.Request) {
-	var v []*process.Process
-
-	v, err := process.Processes()
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	result := []map[string]string{}
-
-	for _, p := range v {
-		name, _ := p.Name()
-		cmd, _ := p.Cmdline()
-		cmdSlice, _ := p.CmdlineSlice()
-
-		pid := fmt.Sprintf("%v", p.Pid)
-
-		if name == "kelp" {
-			if len(cmdSlice) > 1 && cmdSlice[1] != "serve" {
-				m := make(map[string]string)
-				m["pid"] = pid
-				m["cmd"] = cmd
-				m["name"] = name
-
-				result = append(result, m)
-			}
-		}
-	}
-
-	js, err := json.Marshal(result)
-	if err != nil {
-		log.Fatal(err)
-	}
-
-	w.Write(js)
-}
-
-func runKelp(params ...string) string {
-	return runTool("kelp", params...)
-}
-
-func runTool(tool string, params ...string) string {
-	debug := false
-	if debug {
-		log.Println(tool)
-		for _, v := range params {
-			log.Println(v)
-		}
-	}
-
-	cmd := exec.Command(tool, params...)
-
-	var stdOut bytes.Buffer
-	cmd.Stdout = &stdOut
-
-	var stdErr bytes.Buffer
-	cmd.Stderr = &stdErr
-
-	err := cmd.Run()
-	if err != nil {
-		log.Println(stdErr.String())
-
-		// kill returns an err?  Don't put fatal here unless you test killKelp
-		log.Println(err)
-	}
-
-	return stdOut.String()
 }
